chore: Reorganize VS Code extension
